Refactor KeyExpander into a procedural expandKey
