+++ aespython/cbc_mode.py
@@ -7,7 +7,7 @@ Copyright (c) 2010, Adam Newman http://www.caller9.com/
 Licensed under the MIT license http://www.opensource.org/licenses/mit-license.php
 """
 __all__ = "CBCMode",
-from .mode import Mode
+from . import Mode
 
 class CBCMode(Mode):
     """Perform CBC operation on a block and retain IV information for next operation"""
